Add new shared file for non-API code
