+++ lib_device_control/host/control_host_support.h
@@ -9,6 +9,28 @@
 #include <assert.h>
 #include "control.h"
 #include "control_shared.h"
+#include "control_transport_shared.h"
+
+/* The max USB packet size is 64B (USB 2.0 section 5.5.3),
+ * but larger data transfers are fragmented into several packets.
+ * During testing with full speed USB it has been reported that control transfers
+ * larger than 8kB cause glitches in the audio playback, since most of the transfer
+ * time is taken by the control data, limiting the time left for the audio data.
+*/
+#define USB_TRANSACTION_MAX_BYTES 2048
+#define USB_DATA_MAX_BYTES USB_TRANSACTION_MAX_BYTES
+
+// hard limit of 256 bytes for xSCOPE uploads
+#define XSCOPE_UPLOAD_MAX_BYTES (XSCOPE_UPLOAD_MAX_WORDS * 4)
+#define XSCOPE_UPLOAD_MAX_WORDS 64
+// subtract the header size from the total upload size
+#define XSCOPE_DATA_MAX_BYTES (XSCOPE_UPLOAD_MAX_BYTES - 4)
+
+#define I2C_TRANSACTION_MAX_BYTES 256
+#define I2C_DATA_MAX_BYTES (I2C_TRANSACTION_MAX_BYTES - 3)
+
+#define SPI_TRANSACTION_MAX_BYTES 256
+#define SPI_DATA_MAX_BYTES (SPI_TRANSACTION_MAX_BYTES - 3)
 
 #if USE_XSCOPE
